+++ CHANGELOG.md
@@ -11,9 +11,12 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 - Added HeaderLayout and -HeaderEmphasis parameters to Set-PnPWeb
 - Support to specify lcid for Export-PnPTaxonomy for a particular termset
+- Added support in the Navigation cmdlets to manage the site footer on modern sites.
 
 ### Changed
 
+### Deprecated
+- Removed support for the Template Gallery as the gallery itself is not online anymore.
 
 ## [3.8.1904.0]